requirements.txt
pydantic
python-bitcoinlib
python-dotenv
neo4j
bitcoin
base58
pycryptodome
loguru
psycopg2-binary
pymgclient
sqlalchemy
confluent_kafka
loguru
mgp